lang: Implement lossless strings
This commit backs the `Value::String` type with an array of `u8`s
instead of `std::string::String`. `std::string::String`s must be utf-8.
This is a problem for on-disk btrfs "strings", as those strings can be
an arbitrary array of non-nul bytes.

After this commit, we will have lossless strings, where when printed,
each byte will be escaped with
https://doc.rust-lang.org/stable/std/ascii/fn.escape_default.html .

See the above documentation for escaping rules.

Note that we choose to implement lossless strings instead of treating
on-disk strings as `Value::Array` of u8 bytes because most of the time
the on-disk strings are still human readable. And since this language is
tailored to btrfs, this is a reasonable tradeoff.